# Improve font performance in FamilyCollection.LookupFamily