@dipeshmsft 2023-06-09
@ThomasGoulet73 this is a good perf improvement. Thanks for your constant efforts.